move utils tests to test_utils from test_basic
